총 평점 계산 추가, 명세표 insert 하는 중

+++ src/com/comvision/artBridge/grade/model/vo/Grade.java
@@ -7,7 +7,7 @@ public class Grade implements Serializable{
 	private int greade_no;
 	private int board_no;
 	private int member_no;
-	private int grade;
+	private double grade;
 	private String grade_content;
 	private Date grade_date;
 	private String nick_name;
@@ -16,7 +16,7 @@ public class Grade implements Serializable{
 		
 	}
 
-	public Grade(int greade_no, int board_no, int member_no, int grade, String grade_content, Date grade_date,
+	public Grade(int greade_no, int board_no, int member_no, double grade, String grade_content, Date grade_date,
 			String nick_name) {
 		super();
 		this.greade_no = greade_no;
@@ -40,7 +40,7 @@ public class Grade implements Serializable{
 		return member_no;
 	}
 
-	public int getGrade() {
+	public double getGrade() {
 		return grade;
 	}
 
@@ -68,7 +68,7 @@ public class Grade implements Serializable{
 		this.member_no = member_no;
 	}
 
-	public void setGrade(int grade) {
+	public void setGrade(double grade) {
 		this.grade = grade;
 	}
